Use triggers to update the PersonalBests table
This change avoids having to recompute the whole table every time the
leaderboards are updated, as it is updated implicitly when new replays
are inserted. This reduces the amount of disk IO and drops the runtime
of the database write from ~30s to less than a second.

When bad replays need to be deleted manually, the PersonalBests table
will become out of date. To recover from this a `--fix` option has been
added that will force the table to be recomputed in the same way that it
would have been done previously.