Add command to delete individual replays
This actually just moves the replay into a new table that stores all the
deleted replays (just in case we want to investigate them in the future,
or if I delete the wrong thing by mistake).

Also refactor the script to use sub-parsers instead of arguments for
each operation.

And clear out the personal bests before doing a full recomputation. This
prevents deleted replays with no slower replacement from sticking around.